stable state

+++ contracts/openZeppelin/token/ERC20/BasicToken.sol
@@ -14,7 +14,7 @@ contract BasicToken is ERC20Basic {
 
   mapping(address => uint256) balances;
 
-  uint8 public decimals_;
+  uint256 public decimals_;
   uint256 totalSupply_;
   string public name_;
   string public symbol_;
